just update to latest svn

+++ framework/rbac/migrations/schema-mssql.sql
@@ -9,15 +9,22 @@
  * @since 2.0
  */
 
-drop table [auth_assignment];
-drop table [auth_item_child];
-drop table [auth_item];
-drop table [auth_rule];
+if object_id('[auth_assignment]', 'U') is not null
+    drop table [auth_assignment];
+
+if object_id('[auth_item_child]', 'U') is not null
+    drop table [auth_item_child];
+
+if object_id('[auth_item]', 'U') is not null
+    drop table [auth_item];
+
+if object_id('[auth_rule]', 'U') is not null
+    drop table [auth_rule];
 
 create table [auth_rule]
 (
     [name]  varchar(64) not null,
-    [data]  text,
+    [data]  blob,
     [created_at]           integer,
     [updated_at]           integer,
     primary key ([name])
@@ -29,7 +36,7 @@ create table [auth_item]
    [type]                 smallint not null,
    [description]          text,
    [rule_name]            varchar(64),
-   [data]                 text,
+   [data]                 blob,
    [created_at]           integer,
    [updated_at]           integer,
    primary key ([name]),
